requirements.txt
# Magnum Quality Control System - Python Requirements
# PyTorch ve torchvision (CUDA destekli)
torch>=2.1.0
torchvision>=0.16.0

# Computer Vision
opencv-python>=4.8.0
pillow>=10.0.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.11.0

# Machine Learning
scikit-learn>=1.3.0

# Progress bars
tqdm>=4.66.0

# Plotting
matplotlib>=3.7.0

# Path handling
pathlib2>=2.3.7

# Logging
colorlog>=6.7.0

# Configuration
pyyaml>=6.0

# Web interface (optional)
flask>=2.3.0
flask-cors>=4.0.0

# JSON handling
jsonschema>=4.19.0

# Image processing extras
imageio>=2.31.0
scikit-image>=0.21.0

# Performance monitoring
psutil>=5.9.0

# GPU monitoring
gpustat>=1.1.1
